refactor: make e2e reusable
