+++ chart/values.yaml
@@ -846,10 +846,10 @@ experimental:
 platform:
   # AccessKey defines how vCluster can find the access key used for the platform.
   accessKey:
-    # Value specifies the api key as a regular text value.
+    # Value specifies the access key as a regular text value.
     value: ""
-    # SecretRef defines where to find the platform access key. By default vCluster will search in the following locations in this precedence:
-    # * platform.apiKey.value
+    # SecretRef defines where to find the platform access key. By default, vCluster will search in the following locations in this precedence:
+    # * platform.accessKey.value
     # * environment variable called LICENSE
     # * secret specified under platform.secret.name
     # * secret called "vcluster-platform-api-key" in the vCluster namespace
