+++ test/e2e_isolation_mode/isolation/isolated.go
@@ -1,4 +1,4 @@
-package e2eisolatedmode
+package isolation
 
 import (
 	"context"
