+++ test/e2e_metrics_proxy/metricsproxy/metrics_proxy.go
@@ -1,4 +1,4 @@
-package e2emetricsproxy
+package metricsproxy
 
 import (
 	"context"
