+++ test/e2e_node/e2e_node_suite_test.go
@@ -28,6 +28,7 @@ import (
 	_ "github.com/loft-sh/vcluster/test/e2e/syncer/pvc"
 	_ "github.com/loft-sh/vcluster/test/e2e/syncer/services"
 	_ "github.com/loft-sh/vcluster/test/e2e/webhook"
+	_ "github.com/loft-sh/vcluster/test/e2e_node/node"
 )
 
 var (
