+++ test/e2e_pause_resume/e2e_pause_resume_vcluster_test.go
@@ -15,6 +15,8 @@ import (
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	apiregistrationv1 "k8s.io/kube-aggregator/pkg/apis/apiregistration/v1"
 
+	_ "github.com/loft-sh/vcluster/test/e2e_pause_resume/pauseresume"
+
 	// Enable cloud provider auth
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 	// Register tests
