+++ test/e2e_plugin/plugin/plugin.go
@@ -1,4 +1,4 @@
-package e2eplugin
+package plugin
 
 import (
 	"time"
