+++ test/e2e_rootless/e2e_rootless_mode_suite_test.go
@@ -29,6 +29,7 @@ import (
 	_ "github.com/loft-sh/vcluster/test/e2e/syncer/pvc"
 	_ "github.com/loft-sh/vcluster/test/e2e/syncer/services"
 	_ "github.com/loft-sh/vcluster/test/e2e/webhook"
+	_ "github.com/loft-sh/vcluster/test/e2e_rootless/rootless"
 )
 
 var (
