+++ test/e2e_target_namespace/targetnamespace/target_namespace.go
@@ -1,4 +1,4 @@
-package e2etargetnamespace
+package targetnamespace
 
 import (
 	"context"
